+++ README.md
@@ -25,7 +25,7 @@ pip install -r requirements.txt
 
 then run application
 ```bash
-flask run
+python app.py
 ```
 
 ## Database
