+++ routes/general/__init__.py
@@ -1,0 +1,1 @@
+from .general import general_bp
