+++ routes/search/__init__.py
@@ -1,1 +1,2 @@
 
+from .search import search_bp
